27/3 m3a 7 ou 29 tllil

+++ indexing_elements.c
@@ -6,7 +6,7 @@
 /*   By: rennacir <rennacir@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/18 17:32:04 by rennacir          #+#    #+#             */
-/*   Updated: 2023/03/22 13:11:09 by rennacir         ###   ########.fr       */
+/*   Updated: 2023/03/27 11:45:41 by rennacir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ int	find_max(t_stack **stack_a)
 int	find_min_with_cond(t_stack **stack_a)
 {
 	t_stack *tmp;
-	int i;
+	long i;
 	tmp = *stack_a;
 	i = find_max(stack_a) + 1;
 	while(tmp)
